+++ tests/test_responses.py
@@ -1,7 +1,9 @@
 import anyio
+import datetime as dt
 import os
-
+import time
 import pytest
+from http.cookies import SimpleCookie
 
 from yaa import status
 from yaa.background import BackgroundTask
@@ -244,7 +246,10 @@ def test_file_response_with_chinese_filename(tmpdir, client_factory):
     assert response.headers["content-disposition"] == expected_disposition
 
 
-def test_set_cookie(client_factory):
+def test_set_cookie(client_factory, monkeypatch):
+    mocked_now = dt.datetime(2100, 1, 22, 12, 0, 0, tzinfo=dt.timezone.utc)
+    monkeypatch.setattr(time, "time", lambda: mocked_now.timestamp())
+
     async def app(scope, receive, send):
         response = Response("Hello, world!", media_type="text/plain")
         response.set_cookie(
@@ -263,6 +268,37 @@ def test_set_cookie(client_factory):
     client = client_factory(app)
     response = client.get("/")
     assert response.text == "Hello, world!"
+    assert (
+        response.headers["set-cookie"]
+        == "mycookie=myvalue; Domain=localhost; expires=Fri, 22 Jan 2100 12:00:10 GMT; "
+        "HttpOnly; Max-Age=10; Path=/; SameSite=none; Secure"
+    )
+
+
+@pytest.mark.parametrize(
+    "expires",
+    [
+        pytest.param(
+            dt.datetime(2100, 1, 22, 12, 0, 10, tzinfo=dt.timezone.utc), id="datetime"
+        ),
+        pytest.param("Fri, 22 Jan 2100 12:00:10 GMT", id="str"),
+        pytest.param(10, id="int"),
+    ],
+)
+def test_expires_on_set_cookie(client_factory, monkeypatch, expires):
+    # Mock time used as a reference for `Expires` by stdlib `SimpleCookie`.
+    mocked_now = dt.datetime(2100, 1, 22, 12, 0, 0, tzinfo=dt.timezone.utc)
+    monkeypatch.setattr(time, "time", lambda: mocked_now.timestamp())
+
+    async def app(scope, receive, send):
+        response = Response("Hello, world!", media_type="text/plain")
+        response.set_cookie("mycookie", "myvalue", expires=expires)
+        await response(scope, receive, send)
+
+    client = client_factory(app)
+    response = client.get("/")
+    cookie: SimpleCookie = SimpleCookie(response.headers.get("set-cookie"))
+    assert cookie["mycookie"]["expires"] == "Fri, 22 Jan 2100 12:00:10 GMT"
 
 
 def test_delete_cookie(client_factory):